feat: Add higress provider to ingress2gateway
This commit adds the higress provider to the ingress2gateway package. The provider is imported in the `print.go` file and a new file `storage.go` is added to the `higress` package. The `storage` struct and `OrderedIngressMap` type are defined in the `storage.go` file. The `newResourcesStorage` function initializes a new `storage` instance. The `List` method is added to the `OrderedIngressMap` type to return a list of `networkingv1.Ingress` objects. The `FromMap` method is also added to the `OrderedIngressMap` type to populate the `ingressNames` and `ingressObjects` fields.

Additionally, a new file `resource_reader.go` is added to the `higress` package. The `resourceReader` struct is defined, implementing the `i2gw.CustomResourceReader` interface. The `newResourceReader` function initializes a new `resourceReader` instance. The `readResourcesFromCluster` method reads resources from the cluster and populates the `storage` instance. The `readResourcesFromFile` method reads resources from a file and populates the `storage` instance.

This commit also includes other changes such as test cases, utility functions, and minor fixes from recent commits.

```